docs, state diagram, missing tests

+++ app/src/modules/network/network.c
@@ -162,7 +162,7 @@ static void network_msg_send(struct network_msg *msg)
 }
 
 static void l4_event_handler(struct net_mgmt_event_callback *cb,
-			     uint32_t event,
+			     uint64_t event,
 			     struct net_if *iface)
 {
 	ARG_UNUSED(cb);
@@ -184,7 +184,7 @@ static void l4_event_handler(struct net_mgmt_event_callback *cb,
 }
 
 static void connectivity_event_handler(struct net_mgmt_event_callback *cb,
-				       uint32_t event,
+				       uint64_t event,
 				       struct net_if *iface)
 {
 	ARG_UNUSED(cb);
